Try a new structure for AB testing Jetpack plans
