+++ client/my-sites/plans/jetpack-plans/ab-tests/products-order.ts
@@ -1,0 +1,60 @@
+/**
+ * Internal dependencies
+ */
+import { abtest } from 'calypso/lib/abtest';
+import {
+	PLAN_JETPACK_SECURITY_DAILY,
+	PLAN_JETPACK_SECURITY_DAILY_MONTHLY,
+	PLAN_JETPACK_SECURITY_REALTIME,
+	PLAN_JETPACK_SECURITY_REALTIME_MONTHLY,
+	JETPACK_COMPLETE_PLANS,
+} from 'calypso/lib/plans/constants';
+import {
+	JETPACK_SCAN_PRODUCTS,
+	JETPACK_SEARCH_PRODUCTS,
+	JETPACK_CRM_FREE_PRODUCTS,
+	JETPACK_ANTI_SPAM_PRODUCTS,
+	PRODUCT_JETPACK_BACKUP_DAILY,
+	PRODUCT_JETPACK_BACKUP_DAILY_MONTHLY,
+	PRODUCT_JETPACK_BACKUP_REALTIME,
+	PRODUCT_JETPACK_BACKUP_REALTIME_MONTHLY,
+} from 'calypso/lib/products-values/constants';
+
+const setProductsInPosition = ( slugs: string[], position: number ) =>
+	slugs.reduce( ( map, slug ) => ( { ...map, [ slug ]: position } ), {} );
+
+const backupsFirst = {
+	[ PRODUCT_JETPACK_BACKUP_DAILY ]: 1,
+	[ PRODUCT_JETPACK_BACKUP_DAILY_MONTHLY ]: 1,
+	[ PLAN_JETPACK_SECURITY_DAILY ]: 10,
+	[ PLAN_JETPACK_SECURITY_DAILY_MONTHLY ]: 10,
+	...setProductsInPosition( JETPACK_COMPLETE_PLANS, 20 ),
+	[ PLAN_JETPACK_SECURITY_REALTIME ]: 30,
+	[ PLAN_JETPACK_SECURITY_REALTIME_MONTHLY ]: 30,
+	[ PRODUCT_JETPACK_BACKUP_REALTIME ]: 40,
+	[ PRODUCT_JETPACK_BACKUP_REALTIME_MONTHLY ]: 40,
+	...setProductsInPosition( JETPACK_SCAN_PRODUCTS, 50 ),
+	...setProductsInPosition( JETPACK_ANTI_SPAM_PRODUCTS, 60 ),
+	...setProductsInPosition( JETPACK_SEARCH_PRODUCTS, 70 ),
+	...setProductsInPosition( JETPACK_CRM_FREE_PRODUCTS, 80 ),
+};
+
+const securityFirst = {
+	[ PLAN_JETPACK_SECURITY_DAILY ]: 1,
+	[ PLAN_JETPACK_SECURITY_DAILY_MONTHLY ]: 1,
+	[ PRODUCT_JETPACK_BACKUP_DAILY ]: 10,
+	[ PRODUCT_JETPACK_BACKUP_DAILY_MONTHLY ]: 10,
+	...setProductsInPosition( JETPACK_COMPLETE_PLANS, 20 ),
+	[ PLAN_JETPACK_SECURITY_REALTIME ]: 30,
+	[ PLAN_JETPACK_SECURITY_REALTIME_MONTHLY ]: 30,
+	[ PRODUCT_JETPACK_BACKUP_REALTIME ]: 40,
+	[ PRODUCT_JETPACK_BACKUP_REALTIME_MONTHLY ]: 40,
+	...setProductsInPosition( JETPACK_SCAN_PRODUCTS, 50 ),
+	...setProductsInPosition( JETPACK_ANTI_SPAM_PRODUCTS, 60 ),
+	...setProductsInPosition( JETPACK_SEARCH_PRODUCTS, 70 ),
+	...setProductsInPosition( JETPACK_CRM_FREE_PRODUCTS, 80 ),
+};
+
+export const getProductsOrderVariation = () => {
+	return abtest( 'jetpackProductsOrder' ) === 'securityFirst_test' ? securityFirst : backupsFirst;
+};
