+++ client/my-sites/plans/jetpack-plans/plans-filter-bar/index.tsx
@@ -18,6 +18,7 @@ import { isConnectStore } from 'calypso/my-sites/plans/jetpack-plans/product-gri
 import { TERM_MONTHLY, TERM_ANNUALLY } from 'calypso/lib/plans/constants';
 import useDetectWindowBoundary from '../use-detect-window-boundary';
 import { getHighestAnnualDiscount } from '../utils';
+import { getWithPlanSidesTreatment } from '../ab-tests/plan-sides';
 
 /**
  * Type dependencies
@@ -35,19 +36,17 @@ interface FilterBarProps {
 	duration?: Duration;
 	onDurationChange?: DurationChangeCallback;
 	onProductTypeChange?: ( arg0: ProductType ) => void;
-	withTreatmentVariant: boolean;
 }
 
 type DiscountMessageProps = {
 	primary?: boolean;
-	withTreatmentVariant: boolean;
 };
 
 const CLOUD_MASTERBAR_STICKY = false;
 const CALYPSO_MASTERBAR_HEIGHT = 47;
 const CLOUD_MASTERBAR_HEIGHT = CLOUD_MASTERBAR_STICKY ? 94 : 0;
 
-const DiscountMessage: React.FC< DiscountMessageProps > = ( { primary, withTreatmentVariant } ) => {
+const DiscountMessage: React.FC< DiscountMessageProps > = ( { primary } ) => {
 	const translate = useTranslate();
 	const isMobile: boolean = useMobileBreakpoint();
 
@@ -59,6 +58,8 @@ const DiscountMessage: React.FC< DiscountMessageProps > = ( { primary, withTreat
 	const highestAnnualDiscount = useSelector( ( state ) =>
 		getHighestAnnualDiscount( state, slugsToCheck )
 	);
+
+	const withTreatmentVariant = useSelector( ( state ) => getWithPlanSidesTreatment( state ) );
 
 	if ( ! highestAnnualDiscount ) {
 		return null;
@@ -93,7 +94,6 @@ const PlansFilterBar: React.FC< FilterBarProps > = ( {
 	showDiscountMessage,
 	duration,
 	onDurationChange,
-	withTreatmentVariant,
 } ) => {
 	const translate = useTranslate();
 	const isInConnectStore = useMemo( isConnectStore, [] );
@@ -126,12 +126,7 @@ const PlansFilterBar: React.FC< FilterBarProps > = ( {
 				/>
 				<span className="plans-filter-bar__toggle-on-label">{ translate( 'Bill yearly' ) }</span>
 			</div>
-			{ showDiscountMessage && (
-				<DiscountMessage
-					primary={ durationChecked }
-					withTreatmentVariant={ withTreatmentVariant }
-				/>
-			) }
+			{ showDiscountMessage && <DiscountMessage primary={ durationChecked } /> }
 		</div>
 	);
 };
